mock: create new file with mocked methods
In order to simplify testing, the methods used to mock ec2 calls have
been all put into the same file.
